Split savedata editing template
